pyproject.toml
[tool.poetry]
name = "kodama"
version = "0.1.0"
description = ""
authors = ["Kina <CrafterKina@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyworld = "^0.3.4"
voicevox-client = "^0.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
